group all tests and move app component to the components directory
