+++ src/__tests__/app.test.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { create } from 'react-test-renderer';
-import App from './app';
-import Display from './components/Display';
-import ButtonPanel from './components/ButtonPanel';
+import App from '../components/app';
+import Display from '../components/display';
+import ButtonPanel from '../components/button-panel';
 
 test('renders without crashing', () => {
   const div = document.createElement('div');
