+++ src/index.js
@@ -1,6 +1,6 @@
+import './index.css';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
-import App from './app';
+import App from './components/app';
 
 ReactDOM.render(<App />, document.getElementById('root'));
